completed the delete project "mini use case" -- you can now delete a project from the database via clicking a button
