--- front-end/src/components/DeleteProject.js
@@ -1,8 +1,0 @@
-import React, { Component } from 'react';
-import { Link } from 'react-router-dom';
-import axios from 'axios';
-import ProjectDataService from "../services/ProjectDataService";
-
-export default function DeleteProject() {
-
-}
