+++ front-end/src/pages/ManageProject.js
@@ -1,6 +1,7 @@
-import React, { Component } from 'react';
+import React, {useEffect, useState} from 'react';
 import AddProject from "../components/AddProject";
 import ProjectsList from "../components/ProjectsList";
+
 
 export default function ManageProject() {
     return (
